% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst.R
\name{.validate_rst_input}
\alias{.validate_rst_input}
\title{Validate input for Golosov RST functions}
\usage{
.validate_rst_input(data)
}
\description{
Validate input for Golosov RST functions
}
\keyword{internal}
