% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst.R
\name{rst_classify}
\alias{rst_classify}
\title{Classify observations in Golosov subtypes using RST polygons}
\usage{
rst_classify(data, offset = 5e-16)
}
\arguments{
\item{data}{A data.frame containing numeric columns \code{s1}, \code{s2}, \code{s3}, representing the seat shares of the three largest parties in parliament after the election.}

\item{offset}{Small numeric offset used to avoid polygon-edge ambiguity.}
}
\value{
A data.frame with coordinates and \code{golosov_subtype}.
}
\description{
Classify observations in Golosov subtypes using RST polygons
}
\references{
Golosov, G. V. (2011). Party system classification: A methodological inquiry. \emph{Party Politics}, 17(5), 539--560.

Golosov, G. V. (2014). Towards a classification of the world's democratic party systems, step 2: Placing the units into categories. \emph{Party Politics}, 20(1), 3--15.
}
