% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst.R
\name{rst_coords}
\alias{rst_coords}
\title{Compute Golosov relative size coordinates}
\usage{
rst_coords(data)
}
\arguments{
\item{data}{A data.frame containing numeric columns \code{s1}, \code{s2}, \code{s3}, representing the seat shares of the three largest parties in parliament after the election.}
}
\value{
A data.frame with original columns plus \code{sr}, \code{x}, and \code{y}.
}
\description{
Compute Golosov relative size coordinates
}
\references{
Golosov, G. V. (2011). Party system classification: A methodological inquiry. \emph{Party Politics}, 17(5), 539--560.

Golosov, G. V. (2014). Towards a classification of the world's democratic party systems, step 2: Placing the units into categories. \emph{Party Politics}, 20(1), 3--15.
}
